requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.35.0
ctransformers==0.2.27  # For GGUF inference
torch==2.1.0  # CPU; add --extra-index-url https://download.pytorch.org/whl/cpu for CPU-only
torchaudio==2.1.0
ctranslate2==3.24.0  # Optional for faster inference
whisper==1.1.10  # openai-whisper
tts==0.22.0  # Coqui TTS (includes Silero)
pdfplumber==0.10.3
pymupdf==1.23.8  # Alternative PDF parser
pytesseract==0.3.10
pillow==10.1.0
apscheduler==3.10.4
langchain==0.0.350  # Optional utilities
websockets==12.0
python-multipart==0.0.6  # For file uploads
python-dotenv==1.0.0
pytest==7.4.3
bootstrap-flask==2.4.0  # For Tailwind equiv
chartjs==3.9.1  # JS lib, install via npm if needed
# For Docker: no extras